added functionality to restart button
